deferring: A new package for tracking deferred actions
When we defer planning actions that have direct side-effects -- that is,
resource instance actions -- we need to remember that we did so because
we then need to force deferring planning for any other resources
downstream to preserve the required order of operations. That means we
need to keep track of what we've already deferred so we can check whether
a newly-visited resource instance must also be deferred.

We also need to track deferrals for the purpose of reporting them as part
of the plan presented to the end-user. This change _starts_ to implement
that, but for now the public API is limited only to generating a boolean
signal for whether there are any deferrals at all, because that's the
minimum possible reporting that allows for correct behavior (so the caller
can treat the resulting plan slight differently in the workflow).

In future we'll also want to be able to return placeholder values for
not-yet-expanded resource instances, but that's not fully built out here
yet since for an initial implementation we can cheat and use
cty.DynamicVal as a universal "we know nothing at all" placeholder.

This API is likely to grow and change in forthcoming commits as we
continue wiring these new behaviors into the Terraform modules runtime.