terraform: Evaluator returns DynamicVal for partial-expanded resource
For now this is just an up-top weird exception to minimize the impact on
the code that's runs when not participating in the unknown_instances
experiment.

In future it would be nice to rework this GetResource method to rely on
the ResourceInstanceKeys result even in the known case, since that will
simplify this function and remove a some duplicate code, but our priority
for now is minimizing the possibility that the code for this experiment
can impact users who are not participating in the experiment.