terraform: Callers can force all actions to be deferred
When the modules runtime is being used inside the stacks runtime, it's
possible that a component could refer to another component that has
deferred changes in its plan. In that case, we do still want to plan the
downstream component (to give earlier feedback if there are obvious
problems with it) but we need to force all planned actions to be treated
as deferred so that we preserve the correct dependency ordering across
all objects described in a stack.

This commit only deals with the modules runtime handling that case. We'll
make the stacks runtime use it in a subsequent commit.